+++ Sources/BraintreeDropIn/Public/BraintreeDropIn/BTDropInLocalization.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+
+@interface BTDropInLocalization : NSObject
+
+/// Sets custom translation locales for the Drop-In UI
+///
+/// @param locales An array of locale codes
++ (void)setCustomTranslations:(NSArray<NSString *> *)locales;
+
+@end